multi_pdf_rag_chatbot/requirements.txt
streamlit
langchain
langchainhub
langchain_community
langchain_groq
sentence-transformers
PyPDF2
python-dotenv
faiss-cpu